chore: getIcon move to shared/lib/utils and rename constants in header
