+++ src/shared/lib/utils/getIcon.tsx
