+++ src/shared/lib/utils/index.ts
@@ -1,2 +1,3 @@
+export { getIcon } from './getIcon'
 export { cn } from './merge-cn'
 export { mergeRefs } from './merge-refs'
