+++ src/widgets/header/model/constants/mobile-dropdown-links.ts
@@ -1,5 +1,5 @@
 import { NavLink } from '@/shared/types'
-export const links: NavLink[] = [
+export const LINKS: NavLink[] = [
   { href: '/settings', name: 'Profile Settings' },
   { href: '/statistics', name: 'Statistics' },
   { href: '/favorites', name: 'Favorites' },
