+++ src/widgets/sidebar/ui/desktop-sidebar/desktop-sidebar-item.tsx
@@ -1,11 +1,10 @@
 import { ComponentPropsWithoutRef, forwardRef } from 'react'
 
+import { getIcon } from '@/shared/lib/utils'
 import { NavLink } from '@/shared/types/navlink'
 import clsx from 'clsx'
 import Link from 'next/link'
 import { usePathname } from 'next/navigation'
-
-import { getIcon } from '../../model/utils/getIcon'
 
 type Props = {
   isCollapsed?: boolean
